+++ src/algorithms/signal_source/gnuradio_blocks/ad936x_iio_source.h
@@ -1,6 +1,7 @@
 /*!
  * \file ad936x_iio_source.h
- * \brief A direct IIO custom front-end gnss-sdr signal gnuradio block for the AD936x AD front-end family with special FPGA custom functionalities.
+ * \brief A direct IIO custom front-end gnss-sdr signal gnuradio block for the
+ * AD936x AD front-end family with special FPGA custom functionalities.
  * \author Javier Arribas, jarribas(at)cttc.es
  *
  * -----------------------------------------------------------------------------
@@ -46,9 +47,9 @@ using ad936x_iio_source_sptr = gnss_shared_ptr<ad936x_iio_source>;
 ad936x_iio_source_sptr ad936x_iio_make_source_sptr(
     const std::string &pluto_uri_,
     const std::string &board_type_,
-    long long bandwidth_,
-    long long sample_rate_,
-    long long freq_,
+    int64_t bandwidth_,
+    int64_t sample_rate_,
+    int64_t freq_,
     const std::string &rf_port_select_,
     const std::string &rf_filter,
     const std::string &gain_mode_rx0_,
@@ -57,7 +58,7 @@ ad936x_iio_source_sptr ad936x_iio_make_source_sptr(
     double rf_gain_rx1_,
     bool enable_ch0,
     bool enable_ch1,
-    long long freq_2ch,
+    int64_t freq_2ch,
     bool ppsmode_,
     bool customsamplesize_,
     const std::string &fe_ip_,
@@ -92,9 +93,9 @@ private:
     friend ad936x_iio_source_sptr ad936x_iio_make_source_sptr(
         const std::string &pluto_uri_,
         const std::string &board_type_,
-        long long bandwidth_,
-        long long sample_rate_,
-        long long freq_,
+        int64_t bandwidth_,
+        int64_t sample_rate_,
+        int64_t freq_,
         const std::string &rf_port_select_,
         const std::string &rf_filter,
         const std::string &gain_mode_rx0_,
@@ -103,7 +104,7 @@ private:
         double rf_gain_rx1_,
         bool enable_ch0,
         bool enable_ch1,
-        long long freq_2ch,
+        int64_t freq_2ch,
         bool ppsmode_,
         bool customsamplesize_,
         const std::string &fe_ip_,
@@ -118,9 +119,9 @@ private:
     ad936x_iio_source(
         const std::string &pluto_uri_,
         const std::string &board_type_,
-        long long bandwidth_,
-        long long sample_rate_,
-        long long freq_,
+        int64_t bandwidth_,
+        int64_t sample_rate_,
+        int64_t freq_,
         const std::string &rf_port_select_,
         const std::string &rf_filter,
         const std::string &gain_mode_rx0_,
@@ -129,7 +130,7 @@ private:
         double rf_gain_rx1_,
         bool enable_ch0,
         bool enable_ch1,
-        long long freq_2ch,
+        int64_t freq_2ch,
         bool ppsmode_,
         bool customsamplesize_,
         const std::string &fe_ip_,
@@ -140,7 +141,6 @@ private:
         double lo_attenuation_db_,
         bool high_side_lo_,
         int tx_lo_channel_);
-
 
     void ad9361_channel_demux_to_buffer(ad936x_iio_samples *samples_in, int nchannels, gr_vector_void_star &output_items);
     void ad9361_channel_demux_and_record(ad936x_iio_samples *samples_in, int nchannels, std::vector<std::fstream> *files_out);
